no clobbering of preexisting files

+++ gh_subdir_downl.py
@@ -7,13 +7,11 @@ description:
     created as GitHub doesn't provide archiving, i.e. git-archive, see
     https://help.github.com/articles/can-i-archive-a-repository/
 
-    There is an internal variable specifying via regex the file filter to use
-    to only download the desired files, called 'filetypes'.
 todo:
     1. general addition of error messages
 
 Example usage:
-    gh_subdir_downl.py -u 8leggedunicorn -r Udacity -s p1_stroop_effect
+    gh_subdir_downl.py -u 8leggedunicorn -r Udacity -s p1_stroop_effect -f 'README|.*\.(csv|bib|py|tex)'
 '''
 
 import requests
@@ -25,22 +23,22 @@ import argparse
 __author__ = 'Yigal Weinstein'
 
 parser = argparse.ArgumentParser()
-parser.add_argument("-v", "--verbose", help="increase output verbosity", 
+parser.add_argument("-v", "--verbose", help="increase output verbosity",
         action="store_true")
-parser.add_argument("-u", "--user", type=str, help="specify the GitHub user account", 
+parser.add_argument("-u", "--user", type=str, help="specify the GitHub user account",
         required=True)
-parser.add_argument("-r", "--repository", type=str, help="specify the GitHub user's repository", 
+parser.add_argument("-r", "--repository", type=str, help="specify the GitHub user's repository",
         required=True)
-parser.add_argument("-s", "--subdirectory", type=str, help="specify a subdirectory to download", 
+parser.add_argument("-s", "--subdirectory", type=str, help="specify a subdirectory to download",
         required=False)
 fhelp = '''
-Specify via regex what files to download.  As an example: 
+Specify via regex what files to download.  As an example:
 
 -f 'README|.*\.(csv|bib|py|tex)'
 
 to download only csv, bib, py and tex files.
 '''
-parser.add_argument("-f", "--filetypes", type=str, help = fhelp, 
+parser.add_argument("-f", "--filetypes", type=str, help = fhelp,
         required=False)
 
 args = parser.parse_args()
@@ -70,7 +68,7 @@ def gen_url( user, repo, path='' ):
 
 def ls_subdir_content( url, path ):
     # returns list of objects - files, and directories contained within a
-    # repository or subdirectory 
+    # repository or subdirectory
 
     headers = { 'Accept' : 'application/vnd.github.v3.raw' }
     r = requests.get( url, headers = headers )
@@ -82,14 +80,16 @@ def get_file( url, file_name, path ):
         os.makedirs(path)
     # Function to download a file
     file_path = os.path.join( path, file_name )
-    response = requests.get( url, stream=True )
+    if os.path.isfile( file_path ):
+        print( "{f} already exists on the local file system".format( f = file_path ) )
+    else:
+        response = requests.get( url, stream=True )
+        print( '\nDirect link {url}\nDownloading {s}'.format( url = url, s = file_name ) )
 
-    print( '\nDirect link {url}\nDownloading {s}'.format( url = url, s = file_name ) )
-
-    file_content = requests.get(url).text #.encode('utf-8')
-    with open( file_path , 'w' ) as out_file:
-        out_file.write(file_content)
-    print( '{s} retrieved'.format( s = file_name ) )
+        file_content = requests.get(url).text #.encode('utf-8')
+        with open( file_path , 'w' ) as out_file:
+            out_file.write(file_content)
+        print( '{s} retrieved'.format( s = file_name ) )
 
 def recurs_dl( contents , path ):
     for obj in contents:
